+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [3.6.12](https://github.com/lifinance/sdk/compare/v3.6.11...v3.6.12) (2025-05-05)
+
+
+### Bug Fixes
+
+* disable message signing for custom steps ([#253](https://github.com/lifinance/sdk/issues/253)) ([6a90712](https://github.com/lifinance/sdk/commit/6a90712b6b7289fd8eac71cd62b101fdf4386d2c))
+
+### [3.6.11](https://github.com/lifinance/sdk/compare/v3.6.10...v3.6.11) (2025-04-29)
+
+
+### Bug Fixes
+
+* let the wallet estimate the gas in case of failure ([123860e](https://github.com/lifinance/sdk/commit/123860ebc44b5b35a2c8b65b3f4939d758252486))
+
+### [3.6.10](https://github.com/lifinance/sdk/compare/v3.6.9...v3.6.10) (2025-04-29)
+
+
+### Bug Fixes
+
+* take old gas estimate if we failed to provide a new one ([#251](https://github.com/lifinance/sdk/issues/251)) ([d4c17c1](https://github.com/lifinance/sdk/commit/d4c17c1d58177494e8d606f07220b6459e0fa930))
+
+### [3.6.9](https://github.com/lifinance/sdk/compare/v3.6.8...v3.6.9) (2025-04-28)
+
+
+### Bug Fixes
+
+* **viem:** make EIP-5792 stable ([310a4cc](https://github.com/lifinance/sdk/commit/310a4cc2b613044f68df9d89f478b5fc5644b006))
+
 ### [3.6.8](https://github.com/lifinance/sdk/compare/v3.6.7...v3.6.8) (2025-04-15)